V0.009
Added Number guesser, mainly done but need to implement memory so the program remembers which numbers it guessed.

Made changes to Random Selector, instead of generating a random number and binding that to heads or tails, I found a java class that can give me a random boolean value.
Also added tracking stats, those get printed to the user afterwards.